Rename import_magic to pathmagic
